Implement `SparseMap`
The `SparseMap` struct now contains most of the
code from `SparseSet`. The map contains keys of
unsigned integers and `Copy` values. This way,
values can be integers, or anything else that's
copy such as the unit tuple.

`SparseSet` now wraps `SparseMap` and uses the
unit tuple as every key's value. The variable
that's called `value` in the implementation of
`SparseSet` is really the map's key.