Fix randomized test bug
The two vector of keys and the vector of values
in `sparse_map_insert_random_values` were both
generated independently from each other. Thus,
they had different, random lengths. When zipping
the two, the resulting iterator had the length of
the shorter vector. If the vectors of values was
shorter than the vector of keys, some keys would
not get inserted. When that happened, the loop
asserting that all keys are part of the map failed.