Add `SparseMap` as underlying data strucutre (#8)
* Automatically grow the set's size

- Add a `new` function to create a set with an
  automatic size.
- Add a `set!` macro to initialize sets.
- Remove name ambiguity between capacity and maximum
  value (see issue #5)

* run cargo fmt

* Update exmaple in README

* Cleanup set tests

`sparse_set_contains_value_out_of_bounds` should not
perform any insertions. Insertions of larger values
into sets with a capacity that doesn't initially
fit those values are tested in `sparse_set_grow_as_needed`.

`sparse_set_fit_bounds` is not interesting anymore. Because
the set's capacity is increased automatically, the test
cannot verify any property of the internal bounds. That's a
good thing, because it means that the abstration has improved.

* Implement `SparseMap`

The `SparseMap` struct now contains most of the
code from `SparseSet`. The map contains keys of
unsigned integers and `Copy` values. This way,
values can be integers, or anything else that's
copy such as the unit tuple.

`SparseSet` now wraps `SparseMap` and uses the
unit tuple as every key's value. The variable
that's called `value` in the implementation of
`SparseSet` is really the map's key.

* Fix randomized test bug

The two vector of keys and the vector of values
in `sparse_map_insert_random_values` were both
generated independently from each other. Thus,
they had different, random lengths. When zipping
the two, the resulting iterator had the length of
the shorter vector. If the vectors of values was
shorter than the vector of keys, some keys would
not get inserted. When that happened, the loop
asserting that all keys are part of the map failed.

* Add an example of counting characters in a string

* Cargo fmt

* Test `SpareMap::update`

* Update README with example

* Test that oversized keys panic

Any key must fit a `usize`.

* Switch update arguments

* Fix typo

* Test that the unit tuple trick works

---------

Co-authored-by: Chris Rybicki <chrisr@monada.co>